+++ spec/tokenize_spec.lua
@@ -214,9 +214,9 @@ describe('tokenize', function()
   end)
 
   spec('newlines', function()
-    assert.are.equal(nil, tokenize('a\nb').newlines[1])
+    assert.are.equal(1, tokenize('a\nb').newlines[1])
     assert.are.equal(nil, tokenize('a\nb').newlines[2])
-    assert.are.equal(true, tokenize('a\n\nb').newlines[1])
+    assert.are.equal(2, tokenize('a\n\nb').newlines[1])
     assert.subtable({ 'a', 'b' }, getTokens('a\n\nb'))
   end)